requirements.txt
parlant
python-dotenv
termcolor
langchain
langchain-openai
chromadb
fastapi
uvicorn



